rubocop cleanup in query_routable_controller - move query_routable_controller.rb to controllers/concerns/
